JPA + REST

+++ src/main/java/sk/tuke/gamestudio/service/RatingService.java
@@ -5,6 +5,6 @@ import sk.tuke.gamestudio.entity.Rating;
 public interface RatingService {
     void setRating(Rating rating) throws RatingException;
     double getAverageRating(String game) throws RatingException;
-    double getRating(String game, String player) throws RatingException;
+    int getRating(String game, String player) throws RatingException;
     void reset() throws RatingException;
 }
